reorganize typechain sources and out dir
Also, removed human-readable unused ABIs
