+++ typescript/solver/.gitignore
@@ -1,1 +1,1 @@
-contracts/typechain
+typechain
