--- typescript/solver/contracts/abi/destinationSettler.ts
@@ -1,3 +1,0 @@
-export default [
-  "function fill(bytes32 orderId, bytes calldata originData, bytes calldata fillerData) external",
-];
