--- typescript/solver/contracts/abi/erc20.ts
@@ -1,3 +1,0 @@
-export default [
-  "function balanceOf(address account) external view returns (uint256)",
-];
