+++ typescript/solver/package.json
@@ -8,7 +8,7 @@
     "dev": "tsx --watch ./index.ts",
     "prettier:check": "prettier . --check",
     "prettier:write": "prettier . --write",
-    "contracts:typegen": "typechain --node16-modules --target ethers-v5 --out-dir contracts/typechain contracts/abi/*.json",
+    "contracts:typegen": "typechain --node16-modules --target ethers-v5 --out-dir typechain **/contracts/*.json",
     "postinstall": "yarn contracts:typegen"
   },
   "repository": {
