+++ typescript/solver/solvers/contracts/erc20.json
